pyproject.toml
[project]
name = "sponana"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = ["drake>=1.22.0", "manipulation>=2023.11.0"]

[project.optional-dependencies]
dev = ["jupyterlab"]
